stopgap - continuing yaml wrapper
