+++ Razor/src/Razor/IO/YamlWrapper.cpp
@@ -4,6 +4,150 @@
 #include <vector>
 #include <sstream>
 #include <exception>
+#include "../Utils/Vector.h"
+#include "../Renderer/Model.h"
+
+namespace YAML
+{
+    template<>
+    struct convert<glm::vec3>
+    {
+        static Node encode(const glm::vec3& rhs)
+        {
+            Node node;
+            node.push_back(rhs.x);
+            node.push_back(rhs.y);
+            node.push_back(rhs.z);
+            node.SetStyle(EmitterStyle::Flow);
+            return node;
+        }
+
+        static bool decode(const Node& node, glm::vec3& rhs)
+        {
+            if (!node.IsSequence() || node.size() != 3)
+                return false;
+
+            rhs.x = node[0].as<float>();
+            rhs.y = node[1].as<float>();
+            rhs.z = node[2].as<float>();
+            return true;
+        }
+    };
+
+    template<>
+    struct convert<glm::vec2>
+    {
+        static Node encode(const glm::vec2& rhs)
+        {
+            Node node;
+            node.push_back(rhs.x);
+            node.push_back(rhs.y);
+            node.SetStyle(EmitterStyle::Flow);
+            return node;
+        }
+
+        static bool decode(const Node& node, glm::vec2& rhs)
+        {
+            if (!node.IsSequence() || node.size() != 2)
+                return false;
+
+            rhs.x = node[0].as<float>();
+            rhs.y = node[1].as<float>();
+            return true;
+        }
+    };
+
+    template<>
+    struct convert<Razor::MeshData::Vertex>
+    {
+        static Node encode(const Razor::MeshData::Vertex& rhs)
+        {
+            Node node;
+            node.push_back(rhs.Position);
+            node.push_back(rhs.Normal);
+            node.push_back(rhs.TexCoords);
+            node.push_back(rhs.Tangent);
+            node.push_back(rhs.Bitangent);
+            node.SetStyle(EmitterStyle::Flow);
+            return node;
+        }
+
+        static bool decode(const Node& node, Razor::MeshData::Vertex& rhs)
+        {
+            if (!node.IsSequence() || node.size() != 5)
+                return false;
+
+            rhs.Position = node[0].as<glm::vec3>();
+            rhs.Normal = node[1].as<glm::vec3>();
+            rhs.TexCoords = node[2].as<glm::vec2>();
+            rhs.Tangent = node[3].as<glm::vec3>();
+            rhs.Bitangent = node[4].as<glm::vec3>();
+            return true;
+        }
+    };
+
+    template<>
+    struct convert<Razor::MeshData::Texture>
+    {
+        static Node encode(const Razor::MeshData::Texture& rhs)
+        {
+            Node node;
+            node.push_back(rhs.Id);
+            node.push_back(rhs.Type);
+            node.push_back(rhs.Path);
+            node.SetStyle(EmitterStyle::Flow);
+            return node;
+        }
+
+        static bool decode(const Node& node, Razor::MeshData::Texture& rhs)
+        {
+            if (!node.IsSequence() || node.size() != 3)
+                return false;
+
+            rhs.Id = node[0].as<unsigned int>();
+            rhs.Type = node[1].as<std::string>();
+            rhs.Path = node[2].as<std::string>();
+            return true;
+        }
+    };
+
+    template<>
+    struct convert<Razor::MeshData>
+    {
+        static Node encode(const Razor::MeshData& rhs)
+        {
+            Node node;
+            node.push_back(rhs.Vertices);
+            node.push_back(rhs.Meshes);
+            node.push_back(rhs.Indices);
+            node.push_back(rhs.Textures);
+            node.push_back(rhs.VAO);
+            node.push_back(rhs.VBO);
+            node.push_back(rhs.EBO);
+            node.push_back(rhs.MaterialId);
+            node.push_back(rhs.bHasIndices);
+            node.SetStyle(EmitterStyle::Flow);
+            return node;
+        }
+
+        static bool decode(const Node& node, Razor::MeshData& rhs)
+        {
+            if (!node.IsSequence() || node.size() != 9)
+                return false;
+
+            rhs.Vertices = node[0].as<std::vector<Razor::MeshData::Vertex>>();
+            rhs.Meshes = node[1].as<std::vector<Razor::MeshData>>();
+            rhs.Indices = node[2].as<std::vector<unsigned int>>();
+            rhs.Textures = node[3].as<std::vector<Razor::MeshData::Texture>>();
+            rhs.VAO = node[4].as<unsigned int>();
+            rhs.VBO = node[5].as<unsigned int>();
+            rhs.EBO = node[6].as<unsigned int>();
+            rhs.MaterialId = node[7].as<unsigned int>();
+            rhs.bHasIndices = node[8].as<bool>();
+            return true;
+        }
+    };
+}
 
 namespace Razor
 {
@@ -244,6 +388,27 @@ namespace Razor
         if (emitter) reinterpret_cast<YamlEmitterImpl*>(emitter)->out << YAML::Value << value;
     }
 
+    void yaml_emitter_value_vec2(YamlEmitter* emitter, const Vector2& vec) {
+        if (emitter) {
+            auto impl = reinterpret_cast<YamlEmitterImpl*>(emitter);
+            impl->out << YAML::Value << YAML::Flow << YAML::BeginSeq << vec.X << vec.Y << YAML::EndSeq;
+        }
+	}
+
+    void yaml_emitter_value_vec3(YamlEmitter* emitter, const Vector3& vec) {
+        if (emitter) {
+            auto impl = reinterpret_cast<YamlEmitterImpl*>(emitter);
+            impl->out << YAML::Value << YAML::Flow << YAML::BeginSeq << vec.X << vec.Y << vec.Z << YAML::EndSeq;
+        }
+	}
+
+    void yaml_emitter_value_vec4(YamlEmitter* emitter, const Vector4& vec) {
+        if (emitter) {
+            auto impl = reinterpret_cast<YamlEmitterImpl*>(emitter);
+            impl->out << YAML::Value << YAML::Flow << YAML::BeginSeq << vec.X << vec.Y << vec.Z << vec.W << YAML::EndSeq;
+        }
+    }
+
     int yaml_emitter_get_string(YamlEmitter* emitter, char* buffer, int bufferSize) {
         if (!emitter) return 0;
         auto impl = reinterpret_cast<YamlEmitterImpl*>(emitter);
@@ -267,4 +432,18 @@ namespace Razor
 		return impl->cache.c_str();
 	}
 
+
+    void yaml_node_push_back(YamlNode* node, void* val) {
+        if (!node || !val) return;
+        auto impl = reinterpret_cast<YamlNodeImpl*>(node);
+        try {
+            impl->node.push_back(*reinterpret_cast<YAML::Node*>(val));
+        }
+        catch (const std::exception& e) {
+            set_error(e.what());
+        }
+        catch (...) {
+            set_error("Unknown error in yaml_node_push_back");
+        }
+	}
 }
