+++ Razor/src/Razor/Scene/SceneSerializer.cpp
@@ -3,163 +3,13 @@
 #include "../Coordinator.h"
 #include "../Component.h"
 #include "../IO/YamlWrapper.h"
-
-namespace YAML
-{
-	template<>
-	struct convert<glm::vec3>
-	{
-		static Node encode(const glm::vec3& rhs)
-		{
-			Node node;
-			node.push_back(rhs.x);
-			node.push_back(rhs.y);
-			node.push_back(rhs.z);
-			node.SetStyle(EmitterStyle::Flow);
-			return node;
-		}
-
-		static bool decode(const Node& node, glm::vec3& rhs)
-		{
-			if (!node.IsSequence() || node.size() != 3)
-				return false;
-
-			rhs.x = node[0].as<float>();
-			rhs.y = node[1].as<float>();
-			rhs.z = node[2].as<float>();
-			return true;
-		}
-	};
-
-	template<>
-	struct convert<glm::vec2>
-	{
-		static Node encode(const glm::vec2& rhs)
-		{
-			Node node;
-			node.push_back(rhs.x);
-			node.push_back(rhs.y);
-			node.SetStyle(EmitterStyle::Flow);
-			return node;
-		}
-
-		static bool decode(const Node& node, glm::vec2& rhs)
-		{
-			if (!node.IsSequence() || node.size() != 2)
-				return false;
-
-			rhs.x = node[0].as<float>();
-			rhs.y = node[1].as<float>();
-			return true;
-		}
-	};
-
-	template<>
-	struct convert<Razor::MeshData::Vertex>
-	{
-		static Node encode(const Razor::MeshData::Vertex& rhs)
-		{
-			Node node;
-			node.push_back(rhs.Position);
-			node.push_back(rhs.Normal);
-			node.push_back(rhs.TexCoords);
-			node.push_back(rhs.Tangent);
-			node.push_back(rhs.Bitangent);
-			node.SetStyle(EmitterStyle::Flow);
-			return node;
-		}
-
-		static bool decode(const Node& node, Razor::MeshData::Vertex& rhs)
-		{
-			if (!node.IsSequence() || node.size() != 5)
-				return false;
-
-			rhs.Position = node[0].as<glm::vec3>();
-			rhs.Normal = node[1].as<glm::vec3>();
-			rhs.TexCoords = node[2].as<glm::vec2>();
-			rhs.Tangent = node[3].as<glm::vec3>();
-			rhs.Bitangent = node[4].as<glm::vec3>();
-			return true;
-		}
-	};
-
-	template<>
-	struct convert<Razor::MeshData::Texture>
-	{
-		static Node encode(const Razor::MeshData::Texture& rhs)
-		{
-			Node node;
-			node.push_back(rhs.Id);
-			node.push_back(rhs.Type);
-			node.push_back(rhs.Path);
-			node.SetStyle(EmitterStyle::Flow);
-			return node;
-		}
-
-		static bool decode(const Node& node, Razor::MeshData::Texture& rhs)
-		{
-			if (!node.IsSequence() || node.size() != 3)
-				return false;
-
-			rhs.Id = node[0].as<unsigned int>();
-			rhs.Type = node[1].as<std::string>();
-			rhs.Path = node[2].as<std::string>();
-			return true;
-		}
-	};
-
-	template<>
-	struct convert<Razor::MeshData>
-	{
-		static Node encode(const Razor::MeshData& rhs)
-		{
-			Node node;
-			node.push_back(rhs.Vertices);
-			node.push_back(rhs.Meshes);
-			node.push_back(rhs.Indices);
-			node.push_back(rhs.Textures);
-			node.push_back(rhs.VAO);
-			node.push_back(rhs.VBO);
-			node.push_back(rhs.EBO);
-			node.push_back(rhs.MaterialId);
-			node.push_back(rhs.bHasIndices);
-			node.SetStyle(EmitterStyle::Flow);
-			return node;
-		}
-
-		static bool decode(const Node& node, Razor::MeshData& rhs)
-		{
-			if (!node.IsSequence() || node.size() != 9)
-				return false;
-
-			rhs.Vertices = node[0].as<std::vector<Razor::MeshData::Vertex>>();
-			rhs.Meshes = node[1].as<std::vector<Razor::MeshData>>();
-			rhs.Indices = node[2].as<std::vector<unsigned int>>();
-			rhs.Textures = node[3].as<std::vector<Razor::MeshData::Texture>>();
-			rhs.VAO = node[4].as<unsigned int>();
-			rhs.VBO = node[5].as<unsigned int>();
-			rhs.EBO = node[6].as<unsigned int>();
-			rhs.MaterialId = node[7].as<unsigned int>();
-			rhs.bHasIndices = node[8].as<bool>();
-			return true;
-		}
-	};
-}
+#include "../Utils/Vector.h"
 
 namespace Razor
 {
-	YAML::Emitter& operator<<(YAML::Emitter& Out, const glm::vec3& Vec)
+	Vector3 ToVec3(const glm::vec3& Vec)
 	{
-		Out << YAML::Flow;
-		Out << YAML::BeginSeq << Vec.x << Vec.y << Vec.z << YAML::EndSeq;
-		return Out;
-	}
-
-	template<typename T>
-	YAML::Emitter& operator<<(YAML::Emitter& Out, const T& rhs)
-	{
-		Out << YAML::convert<T>::encode(rhs);
-		return Out;
+		return Vector3(Vec.x, Vec.y, Vec.z);
 	}
 
 	void SceneSerializer::SerializeEntity(YamlEmitter* Out, Entity InEntity)
@@ -169,14 +19,18 @@ namespace Razor
 		yaml_emitter_value_int32(Out, (uint32_t)InEntity.EntityHandle);
 		if (InEntity.HasComponent<Transform>())
 		{
-			Out << YAML::Key << "Transform";
-			Out << YAML::BeginMap;
+			yaml_emitter_key(Out, "Transform");
+			yaml_emitter_begin_map(Out);
 
 			Transform& EntityTransform = InEntity.GetComponent<Transform>();
-			Out << YAML::Key << "Position" << YAML::Value << EntityTransform.Position;
-			Out << YAML::Key << "Rotation" << YAML::Value << EntityTransform.Rotation;
-			Out << YAML::Key << "Scale" << YAML::Value << EntityTransform.Scale;
-			Out << YAML::EndMap;
+			yaml_emitter_key(Out, "Position");
+			yaml_emitter_value_vec3(Out, ToVec3(EntityTransform.Position));
+			yaml_emitter_key(Out, "Rotation");
+			yaml_emitter_value_vec3(Out, ToVec3(EntityTransform.Rotation));
+			yaml_emitter_key(Out, "Scale");
+			yaml_emitter_value_vec3(Out, ToVec3(EntityTransform.Scale));
+			yaml_emitter_end_map(Out);
+			
 		}
 		//if (InEntity.HasComponent<Mesh>())
 		//{
@@ -190,7 +44,7 @@ namespace Razor
 		//	Out << YAML::EndMap;
 		//}
 
-		Out << YAML::EndMap;
+		yaml_emitter_end_map(Out);
 	}
 
 
@@ -244,24 +98,22 @@ namespace Razor
 			RZ_CORE_WARN("SceneSerializer.Deserialize: Attempted to load scene file that didn't exist called: {0}", OutScene->GetPath());
 			return false;
 		}
-		YAML::Node Data;
-		try
+		YamlNode* Data = yaml_load_file(OutScene->GetPath().c_str());
+		if(!Data)
 		{
-			Data = YAML::LoadFile(OutScene->GetPath());
-		}
-		catch (YAML::ParserException e)
-		{
-			RZ_CORE_ERROR("Failed to load .rzscn file '{0}'\n	{1}", OutScene->GetPath(), e.what());
+			RZ_CORE_ERROR("Failed to load .rzscn file '{0}'\n	{1}", OutScene->GetPath(), yaml_get_last_error());
 			return false;
 		}
 
-		if (!Data["Scene"])
+		if (!yaml_get_child(Data, "Scene"))
 		{
 			RZ_CORE_ERROR("Incomplete .rzscn file missing 'Scene' key");
 			return false;
 		}
 
-		std::string SceneName = Data["Scene"].as<std::string>();
+		std::string SceneName;
+		SceneName.reserve(50000);
+		yaml_as_string(yaml_get_child(Data, "Scene"), &SceneName[0], 50000);
 		RZ_CORE_TRACE("Deserializing scene '{0}'", SceneName);
 
 		auto Entities = Data["Entities"];
