+++ Razor/src/Razor/Utils/Vector.h
@@ -1,0 +1,25 @@
+#pragma once
+
+namespace Razor
+{
+	struct Vector2
+	{
+		float X;
+		float Y;
+	};
+
+	struct Vector3
+	{
+		float X;
+		float Y;
+		float Z;
+	};
+
+	struct Vector4
+	{
+		float X;
+		float Y;
+		float Z;
+		float W;
+	};
+}
